fsck: add exfat_repair_rename_ask() to make code reusable
The renaming strategy is provided in handle_dot_dotdot_filename(),
which can be re-used in other repairs that need to be renamed.

So this commit moves the strategy into exfat_repair_rename_ask()
and rename "dot_name_num" to "invalid_name_num".

Code refinement, no functional changes.
